feat: adding types for user login/registration
