+++ src/react/src/app/types.ts
@@ -37,5 +37,21 @@ export interface DetailState {
 export interface TabState {
     tab: number,
     path: string,
-};
+}
 
+export interface User {
+    username: string,
+    password: string,
+}
+
+export interface RegisterUser extends User {
+    email: string,
+}
+export interface UserState {
+    isLoggedIn: boolean,
+    isAdmin: boolean,
+    user: User | null,
+    message: any,
+    status: 'idle' | 'loading' | 'failed';
+
+}
